requirements.txt
Flask
numpy
astropy
matplotlib
parse
Pillow
requests